requirements-frozen.txt
absl-py==0.5.0
aiodns==1.1.1
aiohttp==3.4.4
alabaster==0.7.12
amqp==2.3.2
appnope==0.1.0
architect==0.5.6
astor==0.7.1
async-timeout==3.0.1
attrs==18.2.0
Babel==2.6.0
billiard==3.5.0.4
bleach==3.0.2
boto==2.45.0
boto3==1.6.23
botocore==1.9.23
cchardet==2.1.1
ccxt==1.13.38
celery==4.1.1
celery-redbeat==0.11.1
certifi==2018.10.15
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
cycler==0.10.0
decorator==4.0.10
defusedxml==0.5.0
dj-database-url==0.4.1
dj-static==0.0.6
Django==2.1.2
django-bmemcached==0.2.3
django-cache-memoize==0.1.1
django-cors-headers==2.3.0
django-debug-toolbar==1.6
django-extensions==1.7.5
django-filter==2.0.0
django-rest-swagger==2.1.2
django-storages==1.5.1
django-toolbelt==0.0.1
django-unixtimestampfield==0.3.9
djangorestframework==3.7.7
docutils==0.14
entrypoints==0.2.3
future==0.16.0
gast==0.2.0
grpcio==1.15.0
gunicorn==19.9.0
h5py==2.8.0
idna==2.6
imagesize==1.1.0
ipykernel==5.1.0
ipyparallel==6.2.2
ipython==5.1.0
ipython-genutils==0.2.0
ipywidgets==7.4.2
itypes==1.1.0
Jinja2==2.10
jmespath==0.9.3
jsonfield==1.0.3
jsonschema==2.6.0
jupyter-client==5.2.3
jupyter-core==4.4.0
Keras==2.2.4
Keras-Applications==1.0.6
Keras-Preprocessing==1.0.5
kiwisolver==1.0.1
kombu==4.2.1
Markdown==3.0.1
MarkupSafe==1.0
matplotlib==3.0.0
mistune==0.8.4
multidict==4.4.2
nbconvert==5.4.0
nbformat==4.4.0
nose==1.3.7
notebook==5.7.0
numpy==1.14.4
oauthlib==2.1.0
openapi-codec==1.3.2
packaging==18.0
pandas==0.23.4
pandocfilters==1.4.2
pexpect==4.6.0
pickleshare==0.7.5
prometheus-client==0.4.2
prompt-toolkit==1.0.15
protobuf==3.6.1
psycopg2==2.7.5
ptyprocess==0.6.0
pycares==2.3.0
Pygments==2.2.0
pyparsing==2.2.2
python-binary-memcached==0.26.1
python-dateutil==2.6.0
python-memcached==1.59
python-telegram-bot==10.1.0
pytz==2016.10
PyYAML>=4.2b1
pyzmq==17.1.2
qtconsole==4.4.2
redis==2.10.6
requests==2.18.4
requests-oauthlib==0.8.0
s3transfer==0.1.13
schedule==0.4.3
scipy==1.1.0
Send2Trash==1.5.0
simplegeneric==0.8.1
simplejson==3.11.1
six==1.11.0
snowballstemmer==1.2.1
Sphinx==1.8.1
sphinxcontrib-websupport==1.1.0
sqlparse==0.2.4
static3==0.7.0
TA-Lib==0.4.17
tensorboard==1.10.0
tensorflow==1.10.1
termcolor==1.1.0
terminado==0.8.1
testpath==0.4.2
toolz==0.9.0
tornado==5.1.1
traitlets==4.3.2
uritemplate==3.0.0
urllib3>=1.23
vine==1.1.4
waitress==1.0.2
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.14.1
whitenoise==3.3.1
widgetsnbextension==3.4.2
yarl==1.1.0
